+++ docs/changelog.md
@@ -12,6 +12,7 @@ Click on the version in the following list to see all changes since the previous
 ### [Unreleased]
 
 * docs: add version headings for older releases of Scrum Poker
+* test: use Mockito version 3.1.0 for mocks in tests
 
 ### [4.5.0] - 2019-10-11