+++ docs/changelog.md
@@ -18,6 +18,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: display comments for issue on Scrum Poker session page
 * feat: add configuration option to display comments on Scrum Poker session page
 * docs: create online documentation for configuration options
+* test: fix test for default option during upgrade task
 
 ### [4.5.0] - 2019-10-11